+++ src/hooks/useGCLogin.ts
@@ -1,6 +1,5 @@
 import { apiGet } from '../utils/api';
 import Cookies from 'js-cookie';
-import { GRAND_CENTRAL_TOKEN_COOKIE } from '../constants/cookie';
 import axios from 'axios';
 
 const authApi = axios.create();
@@ -9,9 +8,10 @@ export type GCLoginParams = {
   token: string;
   refresh?: string | null;
   gcUrl: string;
+  sessionCookieName: string;
 };
 export default function useGCLogin() {
-  return async ({ token, refresh, gcUrl }: GCLoginParams) => {
+  return async ({ token, refresh, gcUrl, sessionCookieName }: GCLoginParams) => {
     let qs = `?token=${encodeURIComponent(token)}`;
     if (refresh) {
       qs += `&refresh=${encodeURIComponent(refresh)}`;
@@ -19,7 +19,7 @@ export default function useGCLogin() {
     try {
       const res = await apiGet(authApi, `${gcUrl}/api/auth${qs}`, {});
       if (res.success && res.status == 200) {
-        Cookies.set(GRAND_CENTRAL_TOKEN_COOKIE, token);
+        Cookies.set(sessionCookieName, token);
         return true;
       }
       return false;